Updated unpacking queue status for lst files
In the case of lst files, only one event is being read, leading to the situation that the status always showed the unpacking to work on 1/1 events.
I've changed this to be the actual number of fragments created out of the number of fragments calculated from the size of the raw event.