Added ability into TTACHit to have a temperature correction applied - We might want to make this a more general thing where any hit can have a graph passed to it for corrections. - This is the way it's going to be for now though - `TempCorrectedCharge(correction_graph)` returns the raw temp corrected charge - `GetTempCorrectedCharge(cor_graph)` returns the kValue divided temp corrected charge - `GetTempCorrectedEnergy(cor_graph)` returns the kvalue divided temp corrected charge which is then calibrated
The energy stored in the transient member is not overwritten. GetEnergy still returns this value (fat peaks). I don't think it's worth the effort at this point to figure this out. It is up to the user to provide the correct `TGraph*` for that hit at run-time to do the correction. Anything else gets complicated to maintain on our end.