Fixing #1124
Changed GMarker to not store independent x/y coordinates and bins but
use values from the TLines instead.

This means that changing the TLines (by dragging them on the canvas)
changes the GMarker values now, e.g. when fitting a histogram. It also
means that one has to provide the histogram the marker refers to either
via the constructor, or afterwards via the GMarker::SetHist function.
Failure to do so results in a segmentation violation if e.g. asking for
the position of the marker in bins (these functions do not yet check if
the histogram pointer has been set).