Added selection of peak shape to TBGSubtraction (#1197)
The user can now select between TGauss, TRWPeak, TABPeak, or TAB3Peak
for fitting in TBGSubtraction. These classes give a better estimate of
the uncertainty in the area of the fit.

- **TBGSubtraction**: Changed from TPeak to TSinglePeak plus TPeakFitter and
  added combo box to select peak type (defaults to TGauss).
- **TPeakFitter**: Added function to reset the init flag (called
whenever the peak type is changed), and a function to print all
parameters on a single line, this function is now being called before
each fit to show the starting parameters.
- **TSinglePeak**: Added function to print all parameters (without
newline), and function to set the centroid (in case the constructor
without the centroid provided has been used).
- **Peak Classes**: Added function to set the centroid. This function is
now used if the centroid is provided in the constructor, but the default
constructor needs this function to be called separately. This is
necessary because the ROOT way of creating new object with runtime
dependent type uses the default constructor without arguments only.