Added check for parameters near their limit
Not sure if this is the proper way to check this, but for now it seems
to work. The way it is done is by comparing the parameters error with
the square root of the corresponding entry on the diagonal of the
covariance matrix. If the absolute of the difference is larger than 0.1
(arbitrarily chosen value) a red warning message about this parameter is
printed (including the name of the parameter and the two values).
If the user passes the new option "retryfit" to TPeakFitter the fit will
be repeated without any limits (using the last fit result as a starting
point).

Also made sure that a fit with the option "q" is actually quiet (apart
from the printing of the final result).