Fixed issue with y-axis when rebinning
The old commit set the y-axis always to the same absolut range which
doesn't work with the changing bin contents when rebinning. Now we
calculate the ratio of where we are zoomed in in respect to the total
possible range of the y-axis and use that for the new rebinned axis.