New program to set ODB cycle information from PPG data
This program takes any root file with a TPPG saved in it (e.g. fragment,
analysis, or selector output files), and try to set the ODB cycle
information from the PPG data.

If the ODB information in TPPG already matches the PPG data the program
won't do anything, otherwise it will try to determine the durations of
the stages of the cycle from the PPG data and write the updated TPPG to
file.

There are different colors for the output:
- dark yellow: file was not found.
- dark red: failed to find a TPPG in the file, this might happen for
source runs.
- blue: ODB matches PPG data, file is skipped.
- dark blue: PPG data is empty, most likely a source run, file is
skipped.
- red: ODB doesn't match the PPG data after setting it from the data.
This is a weird state to get in (not sure if it's even possible to get
in this state).
- green: ODB information was successfully set from the PPG data.