Added new class for user settings
This new class (TUserSettings) can be used to read user settings (bool,
int, double, or string) from a text file with the extension ".par" or
from a root file. Parameters are accessible by their names. The format
of the .par-files is "<parameter-name>: <value>", where <parameter-name>
is the name of the parameter (without any white spaces or colons) and
<value> is the value of the parameter. The type will be determined
automatically. Lines starting with '#', or '//', or without a colon are
ignored.
I've also modified TGRSIOptions to read and write these user settings.
TGRSIFrame and TGRSIHelper have been modified to pass those user
settings on to any helper written by users, an example of how to use
them can be seen in the angular correlation helper.
I will also add an example parameter file in the next commit.