Checking if PROOF is available
We check if `root-config --has-proof` returns `yes`. If it does,
we add `-lProof` to the list of libraries, and add grsiproof.o as a
target. If not we remove `TGRSIProof` from the list of directories we
lopp over.

This makes the normal `make` command work again on macOS with ROOT
6.32.2 installed. Haven't had a chance to test this on linux with 6.32
or older versions of ROOT (which still have PROOF).