Fixed bug in helper compilation
Using std::stringstream::clear only clears the flags of the object, not
the underlying std::string, so the command executed when "linking" was
actually repeating the compile step as well.
Also improved this to not have linker flags in the compile step and to
properly include the parser libraries shared object libraries. For macOS
an extra include path to /opt/local/include was added (to find
vdt/vdtMath.h).