Added check for nullptrs
We now check if the filename for stdout or stderr are nullptrs. If they
are, we do not redirect stdout or stderr, respectively. This allows
redirecting only one of them at a time. Using nullptrs seems to have
worked before as well, but not sure why, and this now makes it clear?