Made GH1D more interactive
Sofar we've used GCanvas to allow adding new interactive features for
drawing histograms (like moving the range with arrow keys). This has the
benefit of allowing this for all types of histograms drawn (same keys
for ROOT-native 1D histograms and GRSISort specific GH1D histograms).
But it also meant we would have to duplicate code to do this whenever
another part of the program needed that interactiveness (like it was
partially done in TBGSubtraction).

I've now added functionality (specifically adding regions to histograms)
to the GH1D class directly and made it such that the class connects
itself to the pad it is being drawn on to interact with any user actions
in that pad.

This was mainly done for the TSourceCalibration right now, but it could
probably also be used for a revamped TBGSubtraction where the regions
are used to set gates and backgrounds.