Trying to resolve issue where points can't be removed
While updating TSourceCalibration I noticed an issue with removing
points from the graph using right-clicks. To try and improve this I
added a lot more verbosity (if enabled via
TCalibrationGraphSet::VerboseLevel) and instead of having the
TCalibrationGraphSet try and determine the position of the interaction,
the TCalibrationGraph itself (which is the object the user interacts
with) determines the position and sends it to the TCalibrationGraphSet.
The TCalibrationGraphSet functions now also emit signals
("RemovePoint(Int_t, Int_t)", and "RemoveResidualPoint(Int_t,Int_t)"
which should allow other classes to pick up on this and update
themselves (like TSourceCalibration removing indicators that a peak is
being used).