Rework of TSourceCalibration
Completely changed the layout and progress to simplify using it.

Instead of first having only the peaks and a "calibration" of a single
source and only seeing the combined data points of all sources (with
residuals) after accepting all source fits (and no way of going back),
we now see all the found peaks and their fits on the left side, while on
the right we see the fit of the combined data points with the residual.
The right side also has a second tab showing the FWHM of all fits which
might help identifying peaks where the fit isn't good or one that are
doublets. It might be better though to move the FWHM graph from the
separate tab to the same canvas as the calibration (maybe as a small pad
of the same width as the calibration that is below/above the calibratio
pad?).

Also implemented a lot of the other improvements from #1409:
- There is now an option to set regions in the histograms to allow
  using higher overall threshold while still finding smaller peaks
  (by having separate regions for strong and weak peaks).
- If possible we list all .sou files that are found in
  $GRSISYS/libraries/TAnalysis/SourceData as possible sources instead of
  just the default Am241, Ba133, Co56, Co60, and Eu152.
- Added "Peak ratio" parameter that determines the maximum number of
  peaks we try to match to the N peaks from the .sou file (limit is
  parameter*N). This is because TSpectrum has the potential of finding
  hundreds of peaks (especially if the threshold is set very low), and
  it takes a very long time to check all possible combinations of those
  peaks and the energies from the .sou file, during which the program is
  completely unresponsive.

Issues that are not completely solved:
- Sometimes(?) not all regions are show after clicking on "Find Peaks"
  or "Find Peaks Fast". If this happens using the 'u'-key or zooming in
  and out again shows all regions again.
- Removing a point from the graph on the right side does not remove the
  corresponding fit of the peak on the left side nor does it change the
  color of the marker for that peak.