Almost working version
Separated functionalities:
- Constructors (of TChannelTab and TSourceTab) now only create the
  graphical interfaces, they don't compute or draw anything.
- The FindPeaks functions only find the peaks, but they don't draw the
  histograms.
- New Draw functions do the job of drawing the projections on the right
  source tabs as well as the calibration and fwhm graphs on the channel
  tabs.

One small issue left is that all the histograms drawn on tabs that
aren't active (meaning not the first source of the first tab), aren't
scaled to the window size. Changing the size of the window fixes that.

Before tackling that issue I will however do the bigger changes planed,
i.e. use the peak positions from TSpectrum to get a rough calibration
and then use that to try and find all peaks for a given source.