Using initial calibration to find peaks
Changed code to use an initial rough calibration (using the results from
TSpectrum directly without fitting the peaks) to find all peaks in the
source files.

Added tab to show that initial calibration, need to check what the
function exactly is, seems to have 0th parameter always set to 1 and
then the offset and the gain?

Doesn't seem to work right away for all channels, for some a second
manual call to find the peaks from the source file seems to work though?

Maybe change it all again and make it so that we only display one
channel at a time? At least for finding peaks and such?