Updated gadd
Copied newest version of hadd, added the loading of libGROOT.so, and
fixed clang-tidy warnings.
This new version allows to use e.g. `gadd -f -e -l Channel` to not
merge the "Channel" objects. Right now the `-l` flag alone could also be
used to set a list of objects to merge, but I will change it so that
"Channel" objects are alwyas excluded, and the -l flag can only be used
to add more objects to be excluded.