Writing SortingDiagnostic to analysis file
This is only done if no fragment tree is being written (otherwise it
will be written to the fragment file anyway). This is a first step to
getting more diagnostics related to the fragment tree into the analysis
file (since we are not writing the fragment tree anymore by default).