First working version of TFragDiagnosticsLoop
This loop creates a separate diagnostics root-file with the same
histograms filled by the AnalyzeDataLoss program (which takes a fragment
tree as input). There might still be mistakes in the logic, especially
where the network packets are involved as the (currently commented out)
error message about a new network packet number that is out of order
(i.e. not greater than or equal to the two stored previous ones)
triggers almost constantly. Might need to investigate if this is due to
the multithreaded frontend GRIFFIN is using nowadays?

Still need to check the CMakeLists.txt file to make this work with cmake
as well.