Memory leaks fixes
Added custom destructors to some classes to delete member pointers.
Main improvement was to add "a" to the call to TDetector::Clear in the
TAnalysisWriteLoop (to delete all hits and not just clear the hit
vector). Changed the definitely lost from 75,145,952 bytes to 11,287,216
bytes, but might be adding errors? Have to double check this.