EfficiencyHelper changed to SummingCorrectionsHelper
The new helper can not only be used to create histograms to determine
summing corrections when calculating efficiencies, but also for summing
corrections for branching ratios. In that case the gamma energy to be
gated on has to be provided, see the SummingCorrections.par file for an
example on how to do this.

Also some small changes (updated comments, or replacing `` with $()).